contracts/ubi: forwardTransfer() - to access other arbitrary tokens.
The EIP223 approach seems overly convoluted to me, and attempts to
account for eventualities that are not "normal process".

Also, its author continuosly pushes it as a "standard", although
it has *not* gone through the standartisation process; in fact, it
_can't_, at least not yet, since even ERC20 isn't yet standartised
(and should be called EIP20).

The idea with EIP223 is that compliant tokens will protect users by
disallowing arbitrary assignment of other compliant tokens. This,
however, does not protect against assignment of non-compliant
tokens.

Instead of ballooning the contract with extraneous logic that can't
be guaranteed to trigger in _all_ cases, provide an interface to
access external "ERC20" tokens.