contracts/dispatcher: simplify initialisation.
Different from Daniel's dispatcher, we do know the contract that
is to be delegated-to, - it's oobiqoo's oo library, the latest
known stable version.

Therefore, there's no need to put the constructor in-code, only
to be skipped on every invocation after the first.

The question of how to get this "latest known stable library
deployment" remains unanswered.