oo: remove `mint(amt)` + `mint()` impl-n + test skeleton.
`mint()` with an argument makes accounting for "last minted time"
more difficult.

(If controlled dilution is found to be a useful feature, it can be
added later on.)

I couldn't find a simple way to set a storage variable, so to
implement the minimal test for `transfer()`, `mint()` had to be
implemented.

The test cases are currently not exhaustive: e.g., access control
is not tested; neither are all the common known edge cases; etc..