oo: fix mint always failing - LOG1 had wrong arg order.
Commit 30a4f47546df6a1dab628be38d4fd505cdca9e05 populated the
*event-minted* definition (and others).

The one place where LOG1 was used directly - in `mint()` impl-n -
had argument order wrong. The first one is memory location, not
event id.

Previosly, when the event id was defined to 0x0, this worked fine by
coincidence. Now, when the event id is a proper 32 bytes, LOG1 was
trying to do memory expansion up to that number - inevitably
consuming all gas.