Revert "contracts/common,oo: introduce per-user approval durations."
This reverts commit 113fba240fb9fef0ac392e2231579c960b355a00.

First off, it bloats the bytecode horribly.

Second, using the "default approve duration" may cost up to
two SLOADs - first for the per-user one, then for the global one
if the former is not set.

Most importantly, though, approval durations are a key piece of
intent, and should not be left out.

The fact that ERC20 approvals are all time-unbound is... unfortunate.