oo: WIP for *collect* (transferFrom()).
Since allowances to/from owner or between third parties are located
in different storage segments, the approval/collection functions
have some ungraceful code repetition.

It can be simplified if the determination of this storage location
is done only once, and the result saved in memory.

This will require some refactoring...