oo: fix collection tests.
The `transferFrom` transaction can't be sent by account A that has
the allowance A->B set.

When performing this collection, the balances of A and B are
changed (as expected). However, the _allowance_ that gets reduced
is that of A->(caller).

If A is the caller, then the allowance to-be-reduced is that of
A->A (A to themselves).

Techically, this _could_ be permitted, since A has an allowance
to spend from A equal to the balance of A. However, this has to
be caught early on in the execution, because otherwise SLOAD and
SSTORE on some-funky-storloc will be performed.

For now, simply forbid this quirky way of transfer.